Evaluation/noisesniffer

See merge request photoholmes/photoholmes!47

+++ src/photoholmes/metrics/IoU_weighted.py
@@ -54,7 +54,7 @@ class IoU_weighted(Metric):
         """
         if preds.shape != target.shape:
             raise ValueError("preds and target must have the same shape")
-
+        target = target.to(torch.int)
         pred_flat = preds.flatten()
         target_flat = target.flatten()
         TPw = torch.sum(pred_flat * target_flat)
